Refactor of accounts app tests
