+++ accounts/tests/test_tokens.py
@@ -1,47 +1,16 @@
-"""accounts app unittests
+"""accounts app unittests for tokens
 
 """
 import base64
 from time import sleep
 
-from django.contrib.auth import get_user_model
-from django.contrib.auth.base_user import BaseUserManager
 from django.test import TestCase
+from django.contrib.auth.base_user import BaseUserManager
 
-from accounts.token import LoginTokenGenerator
+from accounts.tokens import LoginTokenGenerator
+
 
 TEST_EMAIL = 'newvisitor@example.com'
-
-
-class WelcomePageTest(TestCase):
-    """Tests relating to the welcome_page view.
-
-    """
-    def test_uses_welcome_template(self):
-        """The root url should response with the welcome page template.
-
-        """
-        response = self.client.get('/')
-        self.assertTemplateUsed(response, 'accounts/welcome.html')
-
-
-class UserModelTest(TestCase):
-    """Tests for passwordless user model.
-
-    """
-    def test_user_valid_with_only_email(self):
-        """Should not raise if the user model is happy with email only.
-
-        """
-        user = get_user_model()(email=TEST_EMAIL)
-        user.full_clean()
-
-    def test_users_are_authenticated(self):
-        """User objects should be authenticated for views/templates.
-
-        """
-        user = get_user_model()()
-        self.assertTrue(user.is_authenticated())
 
 
 class TokenGeneratorTest(TestCase):
